pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dorsal-arxiv-pdf"
version = "0.1.0"
description = "Dorsal model: arxiv-pdf"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dorsalhub",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "dorsalhub-adapters>=0.2.0"
]

[project.entry-points."dorsal.models"]
arxiv_pdf = "arxiv_pdf"

[tool.hatch.build.targets.wheel]
packages = ["arxiv_pdf"]

[tool.hatch.build.targets.wheel.force-include]
"model_config.toml" = "arxiv_pdf/model_config.toml"
